Move some more stuff to the appropriate class
